adjust layouted
