+++ backend/kotlin-core/src/main/kotlin/br/com/zup/beagle/widget/Widget.kt
@@ -23,20 +23,17 @@ import br.com.zup.beagle.core.Layout
 import br.com.zup.beagle.core.LayoutComponent
 import br.com.zup.beagle.core.Style
 import br.com.zup.beagle.core.StyleComponent
-import br.com.zup.beagle.core.WidgetStyle
-import br.com.zup.beagle.core.WidgetStyleComponent
 
 /**
  * Base of all widgets
  *
  */
 abstract class Widget : StyleComponent, AccessibilityComponent,
-    IdentifierComponent, WidgetStyleComponent, LayoutComponent {
+    IdentifierComponent, LayoutComponent {
 
     override var id: String? = null
     override var style: Style? = null
     override var accessibility: Accessibility? = null
-    override var widgetStyle: WidgetStyle? = null
     override var layout: Layout? = null
 
 }
