+++ backend/kotlin-core/src/test/kotlin/br/com/zup/beagle/ext/SetFlexTest.kt
@@ -16,8 +16,7 @@
 
 package br.com.zup.beagle.ext
 
-import br.com.zup.beagle.core.Style
-import br.com.zup.beagle.core.StyleComponent
+import br.com.zup.beagle.core.Layout
 import br.com.zup.beagle.fake.FlexComponentFake
 import br.com.zup.beagle.widget.core.Size
 import org.junit.jupiter.api.DisplayName
@@ -46,7 +45,7 @@ internal class SetFlexTest {
 
             // Then
             val expected = FlexComponentFake(
-                Style(
+                Layout(
                     flex = br.com.zup.beagle.widget.core.Flex(
                         grow = grow
                     ),
@@ -81,7 +80,7 @@ internal class SetFlexTest {
 
             // Then
             val expected = FlexComponentFake(
-                Style(
+                Layout(
                     flex = br.com.zup.beagle.widget.core.Flex(
                         grow = grow,
                         shrink = shrink
